+++ src/Domain/Repository/TypeRepositoryInterface.php
@@ -206,4 +206,12 @@ interface TypeRepositoryInterface
      * @return Type[]
      */
     public function getFromSheetMeetingRequests(Sheet $sheet, string $locale): array;
+
+    /**
+     * @param Event  $event
+     * @param string $locale
+     *
+     * @return Type[]
+     */
+    public function getTypesAndCategoriesTranslationsByEvent(Event $event, string $locale): array;
 }